LPS-78093 Move WikiCacheThreadLocal to internal.
